feat(9544): remove wip ledger sign tests
